(fix) Fixed endianness to realize linked functions.
The SRE functions, but the event queue must be flushed manually.
This is currently done through a modal (an undocumented feature).
Next steps:
- [ ] Move to a Web Worker (use Transferable objects for messages)
- [X] Ensure get_pos/set_pos are being set

Note that endianness must be little in this port, in contrast
to Julius' normal functioning. I suspect it has to do with
the use of TypedArrays instead of DataViews
to access allocated memory in emscripten,
see src/preamble.js in kripken/emscripten for details.